Fix sitmap.xml generation
Before this commit, sitemap.xml was generated with
an extra "blog" string after runnin the "script" section
in CI. This was due to the --pathprefix="/blog/ option
used on travis.yml to publish on github sites.

After this commit, sitmap.xml is generated correctly,
because the path of the site is taken directly from
metadata.url, without going the url variable which was
modified by the --pathprefix="/blog/ option.