Travis executes eleventy as specified in package.json
Before this commit, travis was not executing eleventy commands
in the same way we were executing them locally.
This was creating discrepancies between how the site was deployed
compared to how we saw locally.

After this commit, also travis uses the --formats flag.
The flag is used to specify to eleventy to not consider other
files than md or njk when transforming.